+++ src/renderers/index.js
@@ -1,0 +1,4 @@
+export {
+  ThreeGraph,
+  create3dGraph,
+} from "./three-graph.js";